refactor(country): move panel coverage rules into shared/
The country-coverage RPC (#7526) has to match headlines, label threats and
cluster incidents exactly the way the country panel does, and server/ may not
import src/app/. Move the rules rather than copy them:

- shared/threat-keyword-classifier.ts — the keyword tables and cascade lifted
  out of src/services/threat-classifier.ts, which keeps the CSS/i18n/LLM halves
  and re-exports classifyByKeyword. Type-equality assertions against src/types
  fail typecheck if the two unions drift.
- shared/country-headline-match.ts — COUNTRY_ALIASES and the first-mention
  precedence rule lifted off CountryIntelManager, which keeps static delegates
  so no call site changed. Header records why this is not the same matcher as
  shared/country-mention.js (brief grounding, all ~194 countries, no precedence).
- shared/country-timeline-events.ts — moved from src/services/, importers and
  its DOM test updated.

Behaviour-preserving: 36/36 classifications identical to the pre-move
implementation across both site variants, country panel DOM suites green,
lint:boundaries clean.
