+++ src/app/country-intel.ts
@@ -16,7 +16,16 @@ import { reverseGeocode } from '@/utils/reverse-geocode';
 import { yieldToMain } from '@/utils/after-paint';
 import { effectivePubDateMs } from '@/services/feed-date';
 import type { CountryCoverageEvent } from '@/services/country-coverage';
-import { reconcileCountryTimelineIncidents } from '@/services/country-timeline-events';
+import { reconcileCountryTimelineIncidents } from '../../shared/country-timeline-events';
+import {
+  COUNTRY_ALIASES,
+  countryTermIndex,
+  escapeRegExp,
+  firstMentionPosition,
+  getCountrySearchTerms,
+  getOtherCountryTerms,
+  isCountryHeadline,
+} from '../../shared/country-headline-match';
 import {
   getCountryAtCoordinates,
   getCountryCentroid,
@@ -1865,85 +1874,30 @@ export class CountryIntelManager implements AppModule {
     BR: { n: 5.3, s: -33.8, e: -34.8, w: -73.9 },
   };
 
-  static COUNTRY_ALIASES: Record<string, string[]> = {
-    IL: ['israel', 'israeli', 'gaza', 'hamas', 'hezbollah', 'netanyahu', 'idf', 'west bank', 'tel aviv', 'jerusalem'],
-    IR: ['iran', 'iranian', 'tehran', 'persian', 'irgc', 'khamenei'],
-    RU: ['russia', 'russian', 'moscow', 'kremlin', 'putin', 'ukraine war'],
-    UA: ['ukraine', 'ukrainian', 'kyiv', 'zelensky', 'zelenskyy'],
-    CN: ['china', 'chinese', 'beijing', 'taiwan strait', 'south china sea', 'xi jinping'],
-    TW: ['taiwan', 'taiwanese', 'taipei'],
-    KP: ['north korea', 'pyongyang', 'kim jong'],
-    KR: ['south korea', 'seoul'],
-    SA: ['saudi', 'riyadh', 'mbs'],
-    SY: ['syria', 'syrian', 'damascus', 'assad'],
-    YE: ['yemen', 'houthi', 'sanaa'],
-    IQ: ['iraq', 'iraqi', 'baghdad'],
-    AF: ['afghanistan', 'afghan', 'kabul', 'taliban'],
-    PK: ['pakistan', 'pakistani', 'islamabad'],
-    IN: ['india', 'indian', 'new delhi', 'modi'],
-    EG: ['egypt', 'egyptian', 'cairo', 'suez'],
-    LB: ['lebanon', 'lebanese', 'beirut'],
-    TR: ['turkey', 'turkish', 'ankara', 'erdogan', 'türkiye'],
-    US: ['united states', 'US', 'u.s.', 'u.s', 'american', 'washington', 'pentagon', 'white house'],
-    GB: ['united kingdom', 'british', 'london', 'uk '],
-    FR: ['france', 'french', 'paris'],
-    BR: ['brazil', 'brazilian', 'brasilia', 'lula', 'bolsonaro'],
-    AE: ['united arab emirates', 'uae', 'emirati', 'dubai', 'abu dhabi'],
-  };
-
-  private static otherCountryTermsCache: Map<string, string[]> = new Map();
+  // The alias table and the matching rules moved to
+  // shared/country-headline-match.ts (#7526) so the country-coverage RPC can
+  // match headlines the way this panel does — server/ may not import src/app/.
+  // These stay as static delegates so no call site changed.
+  static readonly COUNTRY_ALIASES: Record<string, string[]> = COUNTRY_ALIASES;
 
   static escapeRegExp(value: string): string {
-    return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+    return escapeRegExp(value);
   }
 
   static countryTermIndex(text: string, term: string): number {
-    const trimmedTerm = term.trim();
-    if (!trimmedTerm) return -1;
-    // Plain two/three-letter country acronyms are case-sensitive so US does
-    // not match the pronoun "us". Dotted forms such as U.S. are unambiguous
-    // and remain case-insensitive with every full-name and demonym alias.
-    const caseSensitive = /^[A-Z]{2,3}$/.test(trimmedTerm);
-    const matchText = caseSensitive ? text : text.toLowerCase();
-    const matchTerm = caseSensitive ? trimmedTerm : trimmedTerm.toLowerCase();
-    const match = new RegExp(`(^|[^A-Za-z0-9])${CountryIntelManager.escapeRegExp(matchTerm)}(?=$|[^A-Za-z0-9])`).exec(matchText);
-    return match ? match.index + (match[1] ?? '').length : -1;
+    return countryTermIndex(text, term);
   }
 
   static firstMentionPosition(text: string, terms: string[]): number {
-    let earliest = Infinity;
-    for (const term of terms) {
-      const idx = CountryIntelManager.countryTermIndex(text, term);
-      if (idx !== -1 && idx < earliest) earliest = idx;
-    }
-    return earliest;
+    return firstMentionPosition(text, terms);
   }
 
   static getOtherCountryTerms(code: string): string[] {
-    const normalizedCode = code.toUpperCase();
-    const cached = CountryIntelManager.otherCountryTermsCache.get(normalizedCode);
-    if (cached) return cached;
-
-    const dedup = new Set<string>();
-    Object.entries(CountryIntelManager.COUNTRY_ALIASES).forEach(([countryCode, aliases]) => {
-      if (countryCode === normalizedCode) return;
-      aliases.forEach((alias) => {
-        const trimmed = alias.trim();
-        if (trimmed.length > 0) dedup.add(trimmed);
-      });
-    });
-
-    const terms = [...dedup];
-    CountryIntelManager.otherCountryTermsCache.set(normalizedCode, terms);
-    return terms;
+    return getOtherCountryTerms(code);
   }
 
   static isCountryHeadline(title: string, country: string, code: string): boolean {
-    const searchTerms = CountryIntelManager.getCountrySearchTerms(country, code);
-    const otherCountryTerms = CountryIntelManager.getOtherCountryTerms(code);
-    const ourPos = CountryIntelManager.firstMentionPosition(title, searchTerms);
-    const otherPos = CountryIntelManager.firstMentionPosition(title, otherCountryTerms);
-    return ourPos !== Infinity && (otherPos === Infinity || ourPos <= otherPos);
+    return isCountryHeadline(title, country, code);
   }
 
   static resolveCountryName(code: string): string {
@@ -1963,10 +1917,7 @@ export class CountryIntelManager implements AppModule {
   }
 
   static getCountrySearchTerms(country: string, code: string): string[] {
-    const aliases = CountryIntelManager.COUNTRY_ALIASES[code.toUpperCase()];
-    if (aliases) return aliases;
-    if (/^[A-Z]{2}$/i.test(country.trim())) return [];
-    return [country];
+    return getCountrySearchTerms(country, code);
   }
 
   static toFlagEmoji(code: string): string {
